AJAX diagnostic... clarity: 3/4 comfort: 3/4
I think I did OK On this one... we will see... ran out of time at the end and had a synatx issue
come up with the data in the requests, so make sure you figure out the quote situation.
Did not have time to grunt nag and grunt test. Work on that